properly updating and getting the ranking for every level

+++ src/main/java/dao/TrackDao.java
@@ -1,7 +1,7 @@
 package dao;
 
 public interface TrackDao {
-	Integer getRatingByTrackId(int trackId);
+	Float getRatingByTrackId(int trackId);
 	boolean addRating(RatingType rating, int trackId, int userId);
 	boolean save(Track track);
